instaclone/requirements.txt
django
psycopg2-binary
django-allauth
djangorestframework
djangorestframework-simplejwt
boto3
django-storages
pillow
django-cors-headers
python-decouple
django-redis
celery
redis
django-filter
dj-database-url